+++ scglobals/static/tx/changelog.md
@@ -1,10 +1,14 @@
 ## tx handelbars
 
+# 2016-03-21
+* user sees stock card filtered to a location | M
+
 # 2016-03-20
 * BUG: names need to be url escaped
 * user sees drop down of items in the search bar | new reduce query on shipments | M
 * user can select an item and go to its stock card | Sm
 * on the stock card, the user sees a of transactions for that item with the fields date, shipmentid, from, to,expiration, lot num, qty, user | new item specific map query | M
+* user can click on the shipment link of a transaction and go to the shipment page | Sm
 
 ## 2015-12-21
 * sees a date filter | Sm | 2.25h